chore: let callers to configure service name and apm tracer
