+++ pkg/tracegen/distributed.go
@@ -19,23 +19,16 @@ package tracegen
 
 import (
 	"context"
-	"errors"
 
 	apmhttp "go.elastic.co/apm/module/apmhttp/v2"
-	"go.elastic.co/apm/v2"
 	"go.uber.org/zap"
 )
 
 // SendDistributedTrace sends events generated by both APM Go Agent and OTEL library and
 // link them with the same traceID so that they are linked and can be shown in the same trace view
 func SendDistributedTrace(ctx context.Context, cfg *Config, logger *zap.SugaredLogger) error {
-	tracer, err := apm.NewTracer(getUniqueServiceName("service", "intake"), "0.0.1")
-	if err != nil {
-		return errors.New("failed to instantiate apm tracer")
-	}
-
 	cfg.TraceID = newRandomTraceID()
-	txCtx, err := IndexIntakeV2Trace(ctx, *cfg, tracer, logger)
+	txCtx, err := IndexIntakeV2Trace(ctx, *cfg, logger)
 	if err != nil {
 		return err
 	}
@@ -43,5 +36,5 @@ func SendDistributedTrace(ctx context.Context, cfg *Config, logger *zap.SugaredL
 	traceparent := apmhttp.FormatTraceparentHeader(txCtx)
 	tracestate := txCtx.State.String()
 	ctx = SetOTLPTracePropagator(ctx, traceparent, tracestate)
-	return SendOTLPTrace(ctx, *cfg, logger, getUniqueServiceName("service", "otlp"))
+	return SendOTLPTrace(ctx, *cfg, logger, cfg.OTLPServiceName)
 }
